Add more tests for Florence2 model
